+++ gatsby-config.js
@@ -1,9 +1,12 @@
+/**
+ * @type {import('gatsby').GatsbyConfig}
+ */
 module.exports = {
   siteMetadata: {
     siteUrl: 'https://kkm.grankontel.com',
     title: 'KKM site',
     author: 'Thierry Malo',
-    description: 'Zouti pou rédé moun maké é korijé kréyòl maké a yo'
+    description: 'Zouti pou rédé moun maké é korijé kréyòl maké a yo',
   },
   plugins: [
     `gatsby-transformer-json`,
@@ -13,12 +16,24 @@ module.exports = {
         path: `./src/data/`,
       },
     },
-    `gatsby-plugin-image`,
-    `gatsby-plugin-sharp`,
-    `gatsby-transformer-sharp`,
-
     'gatsby-plugin-sass',
-    'gatsby-plugin-react-helmet',
+    'gatsby-plugin-image',
     'gatsby-plugin-sitemap',
+    {
+      resolve: 'gatsby-plugin-manifest',
+      options: {
+        icon: 'src/images/icon.png',
+      },
+    },
+    'gatsby-plugin-sharp',
+    'gatsby-transformer-sharp',
+    {
+      resolve: 'gatsby-source-filesystem',
+      options: {
+        name: 'images',
+        path: './src/images/',
+      },
+      __key: 'images',
+    },
   ],
 }