+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- migrated to gatsby 5
+
 ## [1.0.3] - 2021-08-05
 
 - updated project list